remove uneeded creation of child component
